+++ mitra_api/src/mastodon_api/statuses/views.rs
@@ -86,10 +86,10 @@ use mitra_validators::{
     errors::ValidationError,
     posts::{
         validate_local_post_links,
-        validate_local_reply,
         validate_post_create_data,
         validate_post_mentions,
         validate_post_update_data,
+        validate_reply,
     },
     reactions::validate_reaction_data,
 };
@@ -261,7 +261,7 @@ async fn create_status(
     validate_post_mentions(&post_data.mentions, post_data.visibility)?;
     validate_local_post_links(&post_data.links, post_data.visibility)?;
     if let Some(ref in_reply_to) = maybe_in_reply_to {
-        validate_local_reply(
+        validate_reply(
             in_reply_to,
             current_user.id,
             post_data.visibility,
@@ -485,7 +485,7 @@ async fn edit_status(
     validate_post_mentions(&post_data.mentions, post.visibility)?;
     validate_local_post_links(&post_data.links, post.visibility)?;
     if let Some(ref in_reply_to) = maybe_in_reply_to {
-        validate_local_reply(
+        validate_reply(
             in_reply_to,
             post.author.id,
             post.visibility,