+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Ignore actor's public keys that have different owner.
 - Create reposts for `Article` objects addressed to FEP-1b12 groups.
 - Reject `Update(Object)` activity if `inReplyTo` has changed.
+- Write warning to log if scope widening happens in a remote reply.
 
 ### Fixed